Moves r_packet_transport and a conftest
