+++ benchmarks/transport_montecarlo_packet.py
@@ -13,7 +13,7 @@ import tardis.transport.montecarlo.r_packet as r_packet
 import tardis.transport.montecarlo.utils as utils
 import tardis.transport.frame_transformations as frame_transformations
 import tardis.transport.geometry.calculate_distances as calculate_distances
-import tardis.transport.r_packet_transport as r_packet_transport
+import tardis.transport.montecarlo.r_packet_transport as r_packet_transport
 from benchmarks.benchmark_base import BenchmarkBase
 from tardis.model.geometry.radial1d import NumbaRadial1DGeometry
 from tardis.transport.montecarlo.estimators.radfield_mc_estimators import (
